Move render functions to common/templates/
