+++ packages/core/src/common/index.ts
@@ -1,6 +1,7 @@
 export * from './auth';
 export * from './encoding';
 export * from './hooks';
+export * from './templates';
 export * from './tokens';
 export * from './utils';
 export * from './validation';
