+++ packages/core/src/common/templates/500.debug.html
