+++ packages/core/src/common/templates/index.ts
@@ -1,0 +1,2 @@
+export { renderError } from './render-error';
+export { render, renderToString } from './render';
