+++ packages/core/src/common/templates/render.ts
