+++ packages/core/src/common/utils/index.ts
@@ -3,6 +3,4 @@ export { displayServerURL } from './display-server-url.util';
 export { escapeProp } from './escape-prop';
 export { escape } from './escape';
 export { isInFile } from './is-in-file.util';
-export { renderError } from './render-error.util';
-export { render, renderToString } from './render.util';
 export { streamToBuffer } from './stream-to-buffer';
